Added a first example of login page and homepage
In particular, added:
- base.css, it contains an import for the font and a base body css modifier
- logo.png and logostyle.css, that is useful when using a logo
- homepage.html, that is just a simple homepage with a navbar
-login.html, a basic login form with a couple of redirects if they want to be signed in.

It's really important to add a javascript function that makes some checks on the form. In case no account is found with a particular email, a popup should appear offering to create an account. At that point some of the information can be copied to the signup page